[WIP] Extract ConstraintService from StudentDataService
Next steps:
- Extract NodeStatusService
- Fix tests
