+++ src/assets/wise5/services/studentDataService.ts
@@ -4,9 +4,7 @@ import { Injectable } from '@angular/core';
 import { ConfigService } from './configService';
 import { AnnotationService } from './annotationService';
 import { ProjectService } from './projectService';
-import { UtilService } from './utilService';
 import { HttpClient, HttpParams } from '@angular/common/http';
-import { TagService } from './tagService';
 import { Observable, Subject } from 'rxjs';
 import { DataService } from '../../../app/services/data.service';
 import { ComponentServiceLookupService } from './componentServiceLookupService';
@@ -28,59 +26,6 @@ export class StudentDataService extends DataService {
     annotations: []
   };
   visitedNodesHistory = [];
-  criteriaFunctionNameToFunction = {
-    branchPathTaken: (criteria) => {
-      return this.evaluateBranchPathTakenCriteria(criteria);
-    },
-    isVisible: (criteria) => {
-      return this.evaluateIsVisibleCriteria(criteria);
-    },
-    isVisitable: (criteria) => {
-      return this.evaluateIsVisitableCriteria(criteria);
-    },
-    isVisited: (criteria) => {
-      return this.evaluateIsVisitedCriteria(criteria);
-    },
-    isVisitedAfter: (criteria) => {
-      return this.evaluateIsVisitedAfterCriteria(criteria);
-    },
-    isRevisedAfter: (criteria) => {
-      return this.evaluateIsRevisedAfterCriteria(criteria);
-    },
-    isVisitedAndRevisedAfter: (criteria) => {
-      return this.evaluateIsVisitedAndRevisedAfterCriteria(criteria);
-    },
-    isCompleted: (criteria) => {
-      return this.evaluateIsCompletedCriteria(criteria);
-    },
-    isCorrect: (criteria) => {
-      return this.evaluateIsCorrectCriteria(criteria);
-    },
-    choiceChosen: (criteria) => {
-      return this.evaluateChoiceChosenCriteria(criteria);
-    },
-    score: (criteria) => {
-      return this.evaluateScoreCriteria(criteria);
-    },
-    teacherRemoval: (criteria) => {
-      return this.evaluateTeacherRemovalCriteria(criteria);
-    },
-    usedXSubmits: (criteria) => {
-      return this.evaluateUsedXSubmitsCriteria(criteria);
-    },
-    wroteXNumberOfWords: (criteria) => {
-      return this.evaluateNumberOfWordsWrittenCriteria(criteria);
-    },
-    addXNumberOfNotesOnThisStep: (criteria) => {
-      return this.evaluateAddXNumberOfNotesOnThisStepCriteria(criteria);
-    },
-    fillXNumberOfRows: (criteria) => {
-      return this.evaluateFillXNumberOfRowsCriteria(criteria);
-    },
-    hasTag: (criteria) => {
-      return this.evaluateHasTagCriteria(criteria);
-    }
-  };
 
   private nodeClickLockedSource: Subject<any> = new Subject<any>();
   public nodeClickLocked$: Observable<any> = this.nodeClickLockedSource.asObservable();
@@ -107,9 +52,7 @@ export class StudentDataService extends DataService {
     private componentServiceLookupService: ComponentServiceLookupService,
     private ConfigService: ConfigService,
     private notebookService: NotebookService,
-    protected ProjectService: ProjectService,
-    private TagService: TagService,
-    private UtilService: UtilService
+    protected ProjectService: ProjectService
   ) {
     super(ProjectService);
     this.notebookService.notebookUpdated$.subscribe(() => {
@@ -215,6 +158,7 @@ export class StudentDataService extends DataService {
   }
 
   updateNodeStatuses() {
+    this.constraintService.evaluate();
     this.updateStepNodeStatuses();
     this.updateGroupNodeStatuses();
     this.maxScore = this.getMaxScore();
@@ -260,7 +204,7 @@ export class StudentDataService extends DataService {
     const nodeId = node.id;
     const nodeStatus: any = this.createNodeStatus(nodeId);
     const constraintsForNode = this.getConstraintsThatAffectNode(node);
-    const constraintResults = this.evaluateConstraints(constraintsForNode);
+    const constraintResults = this.constraintService.evaluate(constraintsForNode);
     nodeStatus.isVisible = constraintResults.isVisible;
     nodeStatus.isVisitable = constraintResults.isVisitable;
     this.setNotVisibleIfRequired(nodeId, constraintsForNode, nodeStatus);
@@ -287,21 +231,6 @@ export class StudentDataService extends DataService {
     }
   }
 
-  evaluateConstraints(constraintsForNode) {
-    let isVisible = true;
-    let isVisitable = true;
-    for (const constraintForNode of constraintsForNode) {
-      const tempResult = this.evaluateConstraint(constraintForNode);
-      const action = constraintForNode.action;
-      if (this.isVisibleConstraintAction(action)) {
-        isVisible = isVisible && tempResult;
-      } else if (this.isVisitableConstraintAction(action)) {
-        isVisitable = isVisitable && tempResult;
-      }
-    }
-    return { isVisible: isVisible, isVisitable: isVisitable };
-  }
-
   setNotVisibleIfRequired(nodeId, constraintsForNode, nodeStatus) {
     if (
       constraintsForNode.length == 0 &&
@@ -310,22 +239,6 @@ export class StudentDataService extends DataService {
     ) {
       nodeStatus.isVisible = false;
     }
-  }
-
-  isVisibleConstraintAction(action) {
-    return [
-      'makeThisNodeNotVisible',
-      'makeAllNodesAfterThisNotVisible',
-      'makeAllOtherNodesNotVisible'
-    ].includes(action);
-  }
-
-  isVisitableConstraintAction(action) {
-    return [
-      'makeThisNodeNotVisitable',
-      'makeAllNodesAfterThisNotVisitable',
-      'makeAllOtherNodesNotVisitable'
-    ].includes(action);
   }
 
   updateNodeStatus(nodeId, nodeStatus) {
@@ -367,163 +280,6 @@ export class StudentDataService extends DataService {
       latestComponentStatesForNode.serverSaveTime;
   }
 
-  evaluateConstraint(constraintForNode) {
-    return this.evaluateNodeConstraint(constraintForNode);
-  }
-
-  evaluateNodeConstraint(constraintForNode) {
-    const removalCriteria = constraintForNode.removalCriteria;
-    const removalConditional = constraintForNode.removalConditional;
-    if (removalCriteria == null) {
-      return true;
-    } else {
-      return this.evaluateMultipleRemovalCriteria(removalCriteria, removalConditional);
-    }
-  }
-
-  evaluateMultipleRemovalCriteria(multipleRemovalCriteria, removalConditional) {
-    let result = false;
-    for (let c = 0; c < multipleRemovalCriteria.length; c++) {
-      const singleCriteriaResult = this.evaluateCriteria(multipleRemovalCriteria[c]);
-      if (c === 0) {
-        result = singleCriteriaResult;
-      } else {
-        if (removalConditional === 'any') {
-          result = result || singleCriteriaResult;
-        } else {
-          result = result && singleCriteriaResult;
-        }
-      }
-    }
-    return result;
-  }
-
-  evaluateCriterias(criterias): boolean {
-    for (const criteria of criterias) {
-      if (!this.evaluateCriteria(criteria)) {
-        return false;
-      }
-    }
-    return true;
-  }
-
-  evaluateCriteria(criteria: any): boolean {
-    const criteriaFunction = this.criteriaFunctionNameToFunction[criteria.name];
-    if (criteriaFunction == null) {
-      return true;
-    }
-    return criteriaFunction(criteria);
-  }
-
-  evaluateIsCompletedCriteria(criteria) {
-    return this.isCompleted(criteria.params.nodeId);
-  }
-
-  evaluateIsCorrectCriteria(criteria) {
-    const componentStates = this.getComponentStatesByNodeIdAndComponentId(
-      criteria.params.nodeId,
-      criteria.params.componentId
-    );
-    for (const componentState of componentStates) {
-      if (componentState.studentData.isCorrect) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  evaluateBranchPathTakenCriteria(criteria) {
-    const expectedFromNodeId = criteria.params.fromNodeId;
-    const expectedToNodeId = criteria.params.toNodeId;
-    const branchPathTakenEvents = this.getBranchPathTakenEventsByNodeId(expectedFromNodeId);
-    for (const branchPathTakenEvent of branchPathTakenEvents) {
-      const data = branchPathTakenEvent.data;
-      if (criteria.params.fromNodeId === data.fromNodeId && expectedToNodeId === data.toNodeId) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  evaluateIsVisibleCriteria(criteria) {
-    const nodeStatus = this.getNodeStatusByNodeId(criteria.params.nodeId);
-    return nodeStatus.isVisible;
-  }
-
-  evaluateIsVisitableCriteria(criteria) {
-    const nodeStatus = this.getNodeStatusByNodeId(criteria.params.nodeId);
-    return nodeStatus.isVisitable;
-  }
-
-  evaluateIsVisitedCriteria(criteria) {
-    const events = this.getEvents();
-    for (const event of events) {
-      if (event.nodeId === criteria.params.nodeId && event.event === 'nodeEntered') {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  evaluateIsVisitedAfterCriteria(criteria) {
-    const isVisitedAfterNodeId = criteria.params.isVisitedAfterNodeId;
-    const criteriaCreatedTimestamp = criteria.params.criteriaCreatedTimestamp;
-    const events = this.getEvents();
-    for (const event of events) {
-      if (
-        event.nodeId === isVisitedAfterNodeId &&
-        event.event === 'nodeEntered' &&
-        event.clientSaveTime > criteriaCreatedTimestamp
-      ) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  evaluateIsRevisedAfterCriteria(criteria) {
-    const isRevisedAfterNodeId = criteria.params.isRevisedAfterNodeId;
-    const isRevisedAfterComponentId = criteria.params.isRevisedAfterComponentId;
-    const criteriaCreatedTimestamp = criteria.params.criteriaCreatedTimestamp;
-    const latestComponentStateForComponent = this.getLatestComponentStateByNodeIdAndComponentId(
-      isRevisedAfterNodeId,
-      isRevisedAfterComponentId
-    );
-    return (
-      latestComponentStateForComponent != null &&
-      latestComponentStateForComponent.clientSaveTime > criteriaCreatedTimestamp
-    );
-  }
-
-  evaluateIsVisitedAndRevisedAfterCriteria(criteria) {
-    const isVisitedAfterNodeId = criteria.params.isVisitedAfterNodeId;
-    const isRevisedAfterNodeId = criteria.params.isRevisedAfterNodeId;
-    const isRevisedAfterComponentId = criteria.params.isRevisedAfterComponentId;
-    const criteriaCreatedTimestamp = criteria.params.criteriaCreatedTimestamp;
-    const events = this.getEvents();
-    for (const event of events) {
-      if (
-        this.isVisitedAndRevisedAfter(
-          isVisitedAfterNodeId,
-          isRevisedAfterNodeId,
-          isRevisedAfterComponentId,
-          event,
-          criteriaCreatedTimestamp
-        )
-      ) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  isVisitedAndRevisedAfter(visitNodeId, reviseNodeId, reviseComponentId, event, timestamp) {
-    return (
-      this.isNodeVisitedAfterTimestamp(event, visitNodeId, timestamp) &&
-      this.hasWorkCreatedAfterTimestamp(reviseNodeId, reviseComponentId, event.clientSaveTime)
-    );
-  }
-
   isNodeVisitedAfterTimestamp(event, nodeId, timestamp) {
     return (
       event.nodeId == nodeId && event.event === 'nodeEntered' && event.clientSaveTime > timestamp
@@ -545,34 +301,6 @@ export class StudentDataService extends DataService {
     return branchPathTakenEvents;
   }
 
-  evaluateChoiceChosenCriteria(criteria: any): boolean {
-    const service = this.componentServiceLookupService.getService('MultipleChoice');
-    const latestComponentState = this.getLatestComponentStateByNodeIdAndComponentId(
-      criteria.params.nodeId,
-      criteria.params.componentId
-    );
-    return latestComponentState != null && service.choiceChosen(criteria, latestComponentState);
-  }
-
-  evaluateScoreCriteria(criteria: any): boolean {
-    const params = criteria.params;
-    const scoreType = 'any';
-    const latestScoreAnnotation = this.AnnotationService.getLatestScoreAnnotation(
-      params.nodeId,
-      params.componentId,
-      this.ConfigService.getWorkgroupId(),
-      scoreType
-    );
-    if (latestScoreAnnotation != null) {
-      const scoreValue = this.getScoreValueFromScoreAnnotation(
-        latestScoreAnnotation,
-        params.scoreId
-      );
-      return this.isScoreInExpectedScores(params.scores, scoreValue);
-    }
-    return false;
-  }
-
   getScoreValueFromScoreAnnotation(scoreAnnotation: any, scoreId: string): number {
     if (scoreId == null) {
       return this.AnnotationService.getScoreValueFromScoreAnnotation(scoreAnnotation);
@@ -581,98 +309,11 @@ export class StudentDataService extends DataService {
     }
   }
 
-  evaluateTeacherRemovalCriteria(criteria: any) {
-    return criteria.params.periodId !== this.ConfigService.getPeriodId();
-  }
-
   isScoreInExpectedScores(expectedScores, score) {
     return (
       expectedScores.indexOf(score) != -1 ||
       (score != null && expectedScores.indexOf(score.toString()) != -1)
     );
-  }
-
-  evaluateUsedXSubmitsCriteria(criteria) {
-    const params = criteria.params;
-    return this.getSubmitCount(params.nodeId, params.componentId) >= params.requiredSubmitCount;
-  }
-
-  getSubmitCount(nodeId, componentId) {
-    // counter for manually counting the component states with isSubmit=true
-    let manualSubmitCounter = 0;
-
-    // counter for remembering the highest submitCounter value found in studentData objects
-    let highestSubmitCounter = 0;
-
-    /*
-     * We are counting with two submit counters for backwards compatibility.
-     * Some componentStates only have isSubmit=true and do not keep an
-     * updated submitCounter for the number of submits.
-     */
-    const componentStates = this.getComponentStatesByNodeIdAndComponentId(nodeId, componentId);
-    for (const componentState of componentStates) {
-      if (componentState.isSubmit) {
-        manualSubmitCounter++;
-      }
-      const studentData = componentState.studentData;
-      if (studentData.submitCounter > highestSubmitCounter) {
-        highestSubmitCounter = studentData.submitCounter;
-      }
-    }
-    return Math.max(manualSubmitCounter, highestSubmitCounter);
-  }
-
-  evaluateNumberOfWordsWrittenCriteria(criteria) {
-    const params = criteria.params;
-    const nodeId = params.nodeId;
-    const componentId = params.componentId;
-    const requiredNumberOfWords = params.requiredNumberOfWords;
-    const componentState = this.getLatestComponentStateByNodeIdAndComponentId(nodeId, componentId);
-    if (componentState != null) {
-      const studentData = componentState.studentData;
-      const response = studentData.response;
-      const numberOfWords = this.UtilService.wordCount(response);
-      if (numberOfWords >= requiredNumberOfWords) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  evaluateAddXNumberOfNotesOnThisStepCriteria(criteria) {
-    const params = criteria.params;
-    const nodeId = params.nodeId;
-    const requiredNumberOfNotes = params.requiredNumberOfNotes;
-    try {
-      const notebook = this.notebookService.getNotebookByWorkgroup();
-      const notebookItemsByNodeId = this.getNotebookItemsByNodeId(notebook, nodeId);
-      return notebookItemsByNodeId.length >= requiredNumberOfNotes;
-    } catch (e) {}
-    return false;
-  }
-
-  evaluateFillXNumberOfRowsCriteria(criteria) {
-    const params = criteria.params;
-    const nodeId = params.nodeId;
-    const componentId = params.componentId;
-    const requiredNumberOfFilledRows = params.requiredNumberOfFilledRows;
-    const tableHasHeaderRow = params.tableHasHeaderRow;
-    const requireAllCellsInARowToBeFilled = params.requireAllCellsInARowToBeFilled;
-    const tableService = this.componentServiceLookupService.getService('Table');
-    const componentState = this.getLatestComponentStateByNodeIdAndComponentId(nodeId, componentId);
-    return (
-      componentState != null &&
-      tableService.hasRequiredNumberOfFilledRows(
-        componentState,
-        requiredNumberOfFilledRows,
-        tableHasHeaderRow,
-        requireAllCellsInARowToBeFilled
-      )
-    );
-  }
-
-  evaluateHasTagCriteria(criteria) {
-    return this.TagService.hasTagName(criteria.params.tag);
   }
 
   getNotebookItemsByNodeId(notebook, nodeId) {
